feat: complete additional functions

+++ ft_strlen.c
@@ -1,7 +1,7 @@
 #include "libft.h"
 
 size_t ft_strlen(const char *s) {
-  int slen = 0;
+  size_t slen = 0;
 
   while (s[slen] != '\0') {
     slen++;
